Move useUpdateNodeCounts to new file
